MindSpore/YoloV4/requirements.txt
numpy
opencv_python==4.6.0.66
Pillow==9.3.0
